chore: refactored recent activity
