+++ src/components/widgets/DateTimeWidget.jsx
@@ -1,4 +1,3 @@
-// src/components/widgets/DateTimeWidget.jsx
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { updateCurrentTime, setTimezone, toggleClockMode } from '../../store/slices/widgetSlices/dateTimeWidgetSlice';
