+++ src/components/widgets/NotificationWidget.jsx
@@ -1,4 +1,3 @@
-// src/components/widgets/NotificationWidget.jsx
 import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import {
