+++ src/components/widgets/ProgressWidget.jsx
@@ -1,4 +1,3 @@
-// src/components/widgets/ProgressWidget.jsx
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import {
