+++ src/store/index.js
@@ -4,8 +4,8 @@ import userReducer from './slices/userSlice';
 import tasksReducer from './slices/tasksSlice';
 import weatherReducer from './slices/weatherSlice';
 import notificationsReducer from './slices/widgetSlices/notificationsWidgetSlice';
-import statisticsReducer from './slices/statisticsSlice';
-import recentMessagesReducer from './slices/recentMessagesSlice';
+import statisticsReducer from './slices/widgetSlices/statisticsWidgetSlice';
+import recentMessagesReducer from './slices/widgetSlices/recentMessagesWidgetSlice';
 import recentActivityReducer from './slices/recentActivitySlice';
 import progressReducer from './slices/widgetSlices/progressWidgetSlice';
 import chartReducer from './slices/widgetSlices/chartWidgetSlice';
