+++ src/store/slices/widgetSlices/statisticsWidgetSlice.js
